Update choice group to be wrapped in a sequence.
While our schema is XML standards compliant, not all tools implement the
full standard, and some have issues with the more complex parts of the
schema.

As this change allows a widely used tool (Altova) to work, and does not
change the meaning of the schema nor it's validation, I have included
it.